requirements.txt
locustio
pyzmq
msgpack-python
pyquery
